requirements.txt
python-telegram-bot[job-queue]==20.7
httpx==0.27.0
python-dotenv
